+++ R/flatten_data.R
@@ -52,29 +52,8 @@ flatten_data.flexfile <- function(x, .allocate = TRUE, ...) {
   x_allocated <- attr(x, "allocated")
   x_rolledup <- attr(x, "rolledup")
 
-  # selects all, but provides a quick safety net in case of changes
-  cats <- readflexfile::sfc_mapping %>%
-    dplyr::distinct(.data$standard_category_id, .data$detailed_standard_category_id, .data$direct_or_overhead) %>%
-    dplyr::mutate(detailed_standard_category_id = as.character(.data$detailed_standard_category_id))
-
-  dir_oh <- readflexfile::sfc_mapping %>%
-    dplyr::distinct(.data$standard_category_id, .data$direct_or_overhead)
-
-  # function to join in the sfc category
-  join_sfc <- function(the_table) {
-    the_table %>%
-      dplyr::mutate(detailed_standard_category_id = as.character(.data$detailed_standard_category_id)) %>%
-      dplyr::left_join(cats, by = "detailed_standard_category_id", suffix = c("", "_sfc")) %>%
-      dplyr::mutate(standard_category_id = dplyr::coalesce(.data$standard_category_id,
-                                                           .data$standard_category_id_sfc)) %>%
-      dplyr::select(-.data$direct_or_overhead, -.data$standard_category_id_sfc) %>%
-      dplyr::left_join(dir_oh, by = "standard_category_id")
-  }
-
-  join_sfc_tables <- c("actualcosthourdata", "forecastatcompletioncosthourdata")
-
-  # for the tables where relevant, join in the sfc mappings
-  flexfile_sfc <- purrr::modify_at(x, join_sfc_tables, join_sfc)
+  # join in the sfc mappings
+  flexfile_sfc <- normalize_functional_categories(x, direct_or_oh_mapping = TRUE)
 
   actuals <- flatten_actuals(flexfile_sfc)
   forecasts <- flatten_forecasts(flexfile_sfc)