+++ tools/package_config.R
@@ -109,3 +109,6 @@ flexfile <- read_flexfile(file)
 files <- system.file("extdata", package = "flexample")
 flexfiles <- read_folder(files, read_flexfile)
 
+flexfile %>%
+  normalize_functional_categories()
+